+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 ## Added
 
 - ✨(backend) add semantic search
+- ✨(backend) add analyzers to full-text search
+- ✨(backend) add evaluation command
 - backend application
 - helm chart
 - 🐛(backend) fix missing index creation in 'index/' view